Multiple experiment import
Enabled import of multiple experiments in the UI
Removed the temporary method createTree and put the contents in the init method and a new AddExperiment method in the HyperTreeListPlus class.
In addition the HyperTreeListPlus initiation was corrected, as the agwStyle (6th) arguments were erroneously given to the style (5th) argument.