Add plugins via class decorator
The decorator is ment for the main classes and makes sure that for the class it is applied to a new class is created with the same name. This new class mixes in the plugin classes with the old class, where the plugin classes more or less inherit from the old class.

In the comments another attempt can be found to achieve this using metaclasses. This is very tricky, so this didn't work out.