Simplifying sequencing mapping from Experiment and File and adapting to the new geometric hashing
In the sequencing plugin:

Add a new method in Experiment that selects the data for sequencing mapping and then generates a hash table for each tile.
Remove mapping files from the Experiment class.
Add class variables to init in the File class.
Add a find_sequences method to file that maps the file onto the sequencing data.