Improved Bayes factor calculation
Add extra parameters to the match: hash table distance, number of distances checked and number of tuples checked.
Source tuple points are not used anymore in tuple_match.
Probability of the foreground model is now based on a 2d gaussian function with respect to the distance.

When matching the nearest_neighbour_match, which optimizes the match with linear mapping, is no longer automatically applied. Neither is the function that gets all sequences from the sequencing data.

Update the Example file as well, including a piece of code to import the tile coordinates in the .loc files, instead of loading all the sequencing data. Performing the nearest neighbour match and getting all sequences into the files can now be done in separate for loops. Furthermore the the hash table formation step and the file matching step are now timed.