Implement PolynomialTransform based on numpy.polyfit
skimage.Polynomial transform has problems with large values that occur when calculating the powers of coordinates. This is likely induced when the inverse of a matrix is taken during the calculation. In polyfit the powered coordinates are scaled before finding the least squares solution and afterwards they are inversely scaled. This solves the problem.