Add methods to obtain sequence labelled molecules
get_sequencing_data_for_file method to obtain the section of the sequencing data that is located in the file. Due to the interpolated vertices this will also work with nonlinear transformations.
optimize_sequencing_match_using_visible_sequences so that fluorescently labelled sequences (selected by sequence name) can now be mapped to the all fluorescent spots found in the image.
determine_sequences_at_current_coordinates method to find the closest (fluorescent) sequence for each molecule.
use_sequences_as_molecules method to map the sequence coordinates onto the image and to use these coordinates as molecules.
Change and add sequence related properties in Molecule. The sequence index is now decoupled from the molecule index, so that all sequencing data can be kept in the file.