Analysis datastorage on disk through xarray
.nc files contain selection, coordinates, background and intensity.
pks and traces files are not automatically exported anymore, but they are still imported and automatically converted to .nc files.
Remove Molecule class as it is likely not necessary anymore as we can combine and select all data using xarray.
Update environment to most recent update of everything, including python 3.9.
Separate import and export functions for pks and traces files from File class.