Fix problems with file permission using xarray and netcdf
The netcdf4 engine seems to have a problem with permissions during writing, while the h5netcdf engine seems to be fine.
Furthermore, performing a try: open_dataset in the __getattr__ method is not a good idea, not sure why exactly. This seems to be a problem only when the file is opened repeatedly with short times in between, possibly the file is still open when the next call is made.