Updated parallel processing in Collection
Collection now updates the dictionary of the original objects.
To exclude classes from being updated, add a _do_not_update attribute to the object. This can for example be done in the __getstate__ method. This is useful for example to exclude updates of experiments, so that all files still link to the same experiment.
(Perhaps another way would be to have a singleton experiment class, although in principle there can be multiple experiments in one session)